safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.2_Variable_Statement/S12.2_A8_T8-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected keyword 'in'. Expected ';' after var declaration.
S12.2_A8_T8

PASS Expected parsing failure

TEST COMPLETE
